Create new google-cloud-dns gem package
