+++ google-cloud-core/lib/google/cloud.rb
@@ -110,51 +110,6 @@ module Google
                                                timeout: (timeout || @timeout)
     end
 
-    ##
-    # Creates a new object for connecting to the DNS service.
-    # Each call creates a new connection.
-    #
-    # For more information on connecting to Google Cloud see the [Authentication
-    # Guide](https://googlecloudplatform.github.io/gcloud-ruby/#/docs/guides/authentication).
-    #
-    # @param [String, Array<String>] scope The OAuth 2.0 scopes controlling the
-    #   set of resources and operations that the connection can access. See
-    #   [Using OAuth 2.0 to Access Google
-    #   APIs](https://developers.google.com/identity/protocols/OAuth2).
-    #
-    #   The default scope is:
-    #
-    #   * `https://www.googleapis.com/auth/ndev.clouddns.readwrite`
-    # @param [Integer] retries Number of times to retry requests on server
-    #   error. The default value is `3`. Optional.
-    # @param [Integer] timeout Default timeout to use in requests. Optional.
-    #
-    # @return [Google::Cloud::Dns::Project]
-    #
-    # @example
-    #   require "google/cloud"
-    #
-    #   gcloud = Google::Cloud.new
-    #   dns = gcloud.dns
-    #   zone = dns.zone "example-zone"
-    #   zone.records.each do |record|
-    #     puts record.name
-    #   end
-    #
-    # @example The default scope can be overridden with the `scope` option:
-    #   require "google/cloud"
-    #
-    #   gcloud  = Google::Cloud.new
-    #   dns_readonly = "https://www.googleapis.com/auth/ndev.clouddns.readonly"
-    #   dns = gcloud.dns scope: dns_readonly
-    #
-    def dns scope: nil, retries: nil, timeout: nil
-      require "google/cloud/dns"
-      Google::Cloud.dns @project, @keyfile, scope: scope,
-                                            retries: (retries || @retries),
-                                            timeout: (timeout || @timeout)
-    end
-
     # rubocop:disable Metrics/LineLength
     # Disabled because the readonly scope in the example code is long and we
     # can't shorten it.
