+++ google-cloud-dns/.yardopts
@@ -1,0 +1,9 @@
+--no-private
+--title=Google Cloud DNS
+--markup markdown
+
+./lib/**/*.rb
+-
+README.md
+OVERVIEW.md
+AUTHENTICATION.md
