+++ google-cloud-dns/lib/google/cloud/dns/credentials.rb
