+++ google-cloud-dns/lib/google/cloud/dns/service.rb
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 
-require "google/cloud/version"
 require "google/cloud/errors"
+require "google/cloud/dns/version"
 require "google/apis/dns_v1"
 
 module Google
@@ -37,8 +37,9 @@ module Google
           @project = project
           @credentials = credentials
           @service = API::DnsService.new
-          @service.client_options.application_name    = "gcloud-ruby"
-          @service.client_options.application_version = Google::Cloud::VERSION
+          @service.client_options.application_name    = "google-cloud-dns"
+          @service.client_options.application_version = \
+            Google::Cloud::Dns::VERSION
           @service.request_options.retries = retries || 3
           @service.request_options.timeout_sec = timeout if timeout
           @service.authorization = @credentials.client
